Split common utils into themed submodules
